+++ src/components/common/Splitter.js
